// src/app/flows/fines/fines-mac/fines-mac-offence-details/fines-mac-offence-details-review/fines-mac-offence-details-review-summary/fines-mac-offence-details-review-summary-imposition-table/interfaces/fines-mac-offence-details-review-summary-imposition-table-data.interface.ts
export interface IFinesMacOffenceDetailsReviewSummaryImpositionTableRowTotalData {
  totalAmountImposed: string;
  totalAmountPaid: string;
  totalBalanceRemaining: string;
}

export interface IFinesMacOffenceDetailsReviewSummaryImpositionTableData {
  impositionDescription: string;
  creditor: string;
  amountImposed: string;
  amountPaid: string;
  balanceRemaining: string;
}